Finished exception implementation, de-coupled services and controllers

+++ Backend/api/src/main/java/harmonize/Services/AdminService.java
@@ -3,11 +3,11 @@ package harmonize.Services;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
+import harmonize.ErrorHandling.Exceptions.RoleNotFoundException;
+import harmonize.ErrorHandling.Exceptions.RolePermissionException;
 import harmonize.ErrorHandling.Exceptions.UserNotFoundException;
 import harmonize.ErrorHandling.Exceptions.UsernameTakenException;
 import harmonize.Roles.Role;
@@ -47,50 +47,56 @@ public class AdminService {
         return "\"" + user.getUsername() + "\"" + " was updated to " + "\"" + username + "\"";
     }
 
-    public ResponseEntity<String> deleteUser(int id){
-        User deletedUser = userRepository.findReferenceById(id);
+    @NonNull
+    public String deleteUser(int id){
+        User user = userRepository.findReferenceById(id);
 
-        deletedUser.getRoles().removeAll(deletedUser.getRoles());
+        if(user == null)
+            throw new UserNotFoundException(id);
+
+        user.getRoles().removeAll(user.getRoles());
         userRepository.deleteById(id);
         
-        return new ResponseEntity<>("\"" + deletedUser.getUsername() + "\"" + " has been deleted", HttpStatus.OK);
+        return "\"" + user.getUsername() + "\"" + " has been deleted";
     }
 
-    public ResponseEntity<String> updateRole(int id, String role){
+    @NonNull
+    public String updateRole(int id, String role){
         User user = userRepository.findReferenceById(id);
         Role newRole = roleRepository.findByName(role);
 
         if(user == null)
-            return new ResponseEntity<>("User does not exist", HttpStatus.NOT_FOUND);
+            throw new UserNotFoundException(id);
 
         if(newRole == null)
-            return new ResponseEntity<>("Role does not exist", HttpStatus.NOT_FOUND);
+            throw new RoleNotFoundException(role);
 
         if(user.getRoles().contains(newRole))
-            return new ResponseEntity<>("User already has role", HttpStatus.NOT_FOUND);
+            throw new RolePermissionException(role);
 
         user.getRoles().add(newRole);
         userRepository.save(user);
 
-        return new ResponseEntity<>("\"" + role + "\"" + " has been added to " + "\"" + user.getUsername() + "\"", HttpStatus.OK);
+        return "\"" + role + "\"" + " has been added to " + "\"" + user.getUsername() + "\"";
     }
 
-    public ResponseEntity<String> deleteRole(int id, String role) {
+    @NonNull
+    public String deleteRole(int id, String role) {
         User user = userRepository.findReferenceById(id);
         Role newRole = roleRepository.findByName(role);
 
         if(user == null)
-            return new ResponseEntity<>("User does not exist", HttpStatus.NOT_FOUND);
+            throw new UserNotFoundException(id);
 
-        if(role == null)
-            return new ResponseEntity<>("Role does not exist", HttpStatus.NOT_FOUND);
+        if(newRole == null)
+            throw new RoleNotFoundException(role);
 
         if(!user.getRoles().contains(newRole))
-            return new ResponseEntity<>("User does not have role", HttpStatus.NOT_FOUND);
+            throw new RolePermissionException(role);
 
         user.getRoles().remove(newRole);
         userRepository.save(user);
 
-        return new ResponseEntity<>("\"" + role + "\"" + " has been deleted from " + "\"" + user.getUsername() + "\"", HttpStatus.OK);
+        return "\"" + role + "\"" + " has been deleted from " + "\"" + user.getUsername() + "\"";
     }
 }
